Add DisplayEngine
